lw3-t3-streams Добавил классы Memory

+++ lw3/task3/lib/Stream/Input/Decorator/CDecryptInputStream.h
@@ -31,5 +31,5 @@ public:
 	}
 
 private:
-	uint32_t m_key;
+	uint8_t m_key;
 };
